requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
openpyxl>=3.1.2
pydantic>=2.5.0
python-docx>=1.1.0
